requirements.txt
aiomysql>=0.3.2
Pyrogram>=2.0.106
python-dotenv>=1.2.1